Feat: Clicking run determines displays the new generation.
The game is not using a timer yet. Instead, the generation is updated whenever Run is clicked.
The algorithm now partially implemented. it excludes cells on the boundaries.
Initial grid is random.
As a result, after a few generations we acheive a steady state condition.
Two methods added to GameArea: getAliveNeighbours and runClick which is handler for clicking the run button.

Next: Implement a timer.