Feat: Resume and pause now operational.
The pause button is removed and merged with Run button.
This makes code easier and helps prevent more than one code running at the same time.
Run and pause now occurs in runClick.
A new global variable added timerStatus to determine whether the Run/Pause click was to resume or pause the game.4