Feat: Added two buttons to control grid size.
The user can now choose between two grid sizes: 30x50 and 50x70.
The buttons are added in a new components GridControls which is a child of GameArea.
These buttons have the same handler clearClick in GameArea, which is the same handler for Clear button.
The clearClick method is modified to take in two parameters: a width and a height.
The method stops the timer, resets the board, and changes the size of the board depending on the button clicked.
Clear doesn't change the board size, it just resets it.
Game is now complete. Just need to add comments.